backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
fastf1==3.2.0
python-multipart==0.0.6
pydantic==2.5.3
python-dateutil==2.8.2
pandas>=2.1,<2.2
numpy>=1.26,<2.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-dotenv==1.0.0


